+++ test/Interaction/CleanRegex/Match/Details/Group/NotMatchedGroupTest.php
@@ -2,19 +2,12 @@
 namespace Test\Interaction\TRegx\CleanRegex\Match\Details\Group;
 
 use PHPUnit\Framework\TestCase;
-use Test\Fakes\CleanRegex\Internal\ThrowSubject;
 use Test\Utils\CustomSubjectException;
 use Test\Utils\Functions;
 use TRegx\CleanRegex\Exception\GroupNotMatchedException;
-use TRegx\CleanRegex\Internal\GroupKey\GroupName;
-use TRegx\CleanRegex\Internal\GroupKey\GroupSignature;
-use TRegx\CleanRegex\Internal\Match\Details\Group\GroupDetails;
-use TRegx\CleanRegex\Internal\Match\MatchAll\EagerMatchAllFactory;
-use TRegx\CleanRegex\Internal\Model\Match\RawMatches;
-use TRegx\CleanRegex\Internal\Model\Match\RawMatchesOffset;
 use TRegx\CleanRegex\Internal\Subject;
+use TRegx\CleanRegex\Match\Details\Detail;
 use TRegx\CleanRegex\Match\Details\Group\NotMatchedGroup;
-use TRegx\CleanRegex\Match\Details\NotMatched;
 
 /**
  * @covers \TRegx\CleanRegex\Match\Details\Group\NotMatchedGroup
@@ -105,9 +98,13 @@ class NotMatchedGroupTest extends TestCase
 
     private function matchGroup(Subject $subject = null): NotMatchedGroup
     {
-        return new NotMatchedGroup(
-            $subject ?? new ThrowSubject(),
-            new GroupDetails(new GroupSignature(1, 'first'), new GroupName('first'), new EagerMatchAllFactory(new RawMatchesOffset([]))),
-            new NotMatched(new RawMatches([]), new ThrowSubject()));
+        $string = $subject ?? 'Foo';
+        return pattern($string . '(?<first>first)?')
+            ->match($string)
+            ->first(function (Detail $detail): NotMatchedGroup {
+                /** @var NotMatchedGroup $group */
+                $group = $detail->group('first');
+                return $group;
+            });
     }
 }